+++ sample-apps/react-native/dogfood/src/components/CallControlsComponent.tsx
@@ -7,7 +7,7 @@ import {
   ToggleVideoPublishingButton,
 } from '@stream-io/video-react-native-sdk';
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { StyleSheet, View, ViewStyle } from 'react-native';
 import { appTheme } from '../theme';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Z_INDEX } from '../constants';
@@ -16,24 +16,25 @@ export type CallControlsComponentProps = {
   onChatOpenHandler?: () => void;
   onHangupCallHandler?: () => void;
   unreadCountIndicator?: number;
+  landscape?: boolean;
 };
 
 export const CallControlsComponent = ({
   onChatOpenHandler,
   onHangupCallHandler,
   unreadCountIndicator,
+  landscape,
 }: CallControlsComponentProps) => {
   const { bottom } = useSafeAreaInsets();
+  const landScapeStyles: ViewStyle = {
+    flexDirection: landscape ? 'column-reverse' : 'row',
+    paddingHorizontal: landscape ? 12 : 0,
+    paddingVertical: landscape ? 0 : 12,
+    paddingBottom: landscape ? 0 : Math.max(bottom, appTheme.spacing.lg),
+  };
 
   return (
-    <View
-      style={[
-        styles.callControlsWrapper,
-        {
-          paddingBottom: Math.max(bottom, appTheme.spacing.lg),
-        },
-      ]}
-    >
+    <View style={[styles.callControlsWrapper, landScapeStyles]}>
       <ReactionButton />
       <ChatButton
         onPressHandler={onChatOpenHandler}
@@ -49,9 +50,7 @@ export const CallControlsComponent = ({
 
 const styles = StyleSheet.create({
   callControlsWrapper: {
-    flexDirection: 'row',
     justifyContent: 'space-evenly',
-    paddingVertical: appTheme.spacing.md,
     zIndex: Z_INDEX.IN_FRONT,
     backgroundColor: appTheme.colors.static_grey,
   },