+++ sample-apps/react/egress-composite/src/components/layouts/index.ts
@@ -4,7 +4,11 @@ import { DominantSpeaker, DominantSpeakerScreenShare } from './DominantSpeaker';
 import { PaginatedGrid } from './PaginatedGrid';
 import { Spotlight } from './Spotlight';
 
+export const DEFAULT_LAYOUT: Layout = 'spotlight';
+export const DEFAULT_SCREENSHARE_LAYOUT: ScreenshareLayout = 'spotlight';
+
 export type Layout = 'grid' | 'single-participant' | 'spotlight' | 'mobile';
+// NOTE: should always be a subset of the Layout type
 export type ScreenshareLayout = 'single-participant' | 'spotlight';
 
 export const layoutMap: Record<